requirements.txt
flask~=2.2.3
aiogram~=2.25.1
python-dotenv~=1.0.0
gunicorn~=20.1.0
pymongo~=4.3.3
motor~=3.1.2
bcrypt~=3.2.2
pyjwt~=2.6.0